adiciona lógica de CRUD de unidade

+++ JaPedidos/src/main/java/japedidos/bd/BD.java
@@ -33,17 +33,13 @@ public final class BD {
     
     static public class Categoria {
         public static final String TABLE = "categoria";
-        public static final String[] COLUMNS = {"id", "nome", "descricao"};
+        public static final String[] COLUMNS = {"id", "nome", "descricao"}; // Utilizado em CategoriaTableModel
         
         public static int insert(japedidos.produto.Categoria c) {
-            if (c == null) {
-                throw new NullPointerException();
-            }
-            
-//            if (c.getId() != -1) { // Categoria não pode ser existente
+//            if (c.getId() != -1) { // Categoria não pode ser existente 
 //                throw new IllegalArgumentException();
 //            }
-            if (c.getId() == -1) { // Só cadastra se não houver id
+            if (c != null && c.getId() == -1) { // Só cadastra se não houver id
                 try {
                     Connection conn = BD.getConnection();
                     PreparedStatement insert = conn.prepareStatement(String.format("INSERT INTO %s(nome, descricao) VALUES (?, ?)", TABLE));
@@ -59,15 +55,11 @@ public final class BD {
                 }
             }
             
-            return -1;
+            return 0;
         }
         
         public static int update(japedidos.produto.Categoria c) {
-            if (c == null) {
-                throw new NullPointerException();
-            }
-            
-            if (c.getId() != -1) { // Só atualiza se for categoria existente
+            if (c != null && c.getId() != -1) { // Só atualiza se for categoria existente
                 try {
                     Connection conn = BD.getConnection();
                     PreparedStatement update = conn.prepareStatement(String.format("UPDATE %s SET nome = ?, descricao = ? WHERE id = ?", TABLE));
@@ -86,15 +78,11 @@ public final class BD {
                 }
             }
             
-            return -1;
+            return 0;
         }
         
         public static int delete(japedidos.produto.Categoria c) {
-            if (c == null) {
-                throw new NullPointerException();
-            }
-            
-            if (c.getId() != -1) { // Só deleta se for categoria já existente no banco
+            if (c != null && c.getId() != -1) { // Só deleta se for categoria já existente no banco
                 try {
                     Connection conn = BD.getConnection();
                     PreparedStatement delete = conn.prepareStatement(String.format("DELETE FROM %s WHERE id = ?", TABLE));
@@ -111,7 +99,7 @@ public final class BD {
                 }
             }
             
-            return -1;
+            return 0;
         }
         
         public static japedidos.produto.Categoria selectLast() {
@@ -119,9 +107,11 @@ public final class BD {
                 Connection conn = BD.getConnection();
                 PreparedStatement select = conn.prepareStatement(String.format("SELECT id, nome, descricao FROM %s ORDER BY id DESC LIMIT 1", TABLE));
                 ResultSet rs = select.executeQuery();
-                rs.next();
+                japedidos.produto.Categoria categoria = null;
                 
-                japedidos.produto.Categoria categoria= new japedidos.produto.Categoria(rs.getInt(1), rs.getString(2), rs.getString(3));
+                if (rs.next()) {
+                    categoria = new japedidos.produto.Categoria(rs.getInt(1), rs.getString(2), rs.getString(3));
+                }
                 select.close();
                 conn.close();
                 
@@ -152,9 +142,10 @@ public final class BD {
         }
         
         public static japedidos.produto.Categoria[] parse(ResultSet rs) {
-            ArrayList<japedidos.produto.Categoria> c = new ArrayList<japedidos.produto.Categoria>();
             
             try {
+                
+                ArrayList<japedidos.produto.Categoria> c = new ArrayList<japedidos.produto.Categoria>();
                 while(rs.next()) {
                     final int id = rs.getInt(1);
                     final String nome = rs.getString(2);
@@ -162,15 +153,149 @@ public final class BD {
                     japedidos.produto.Categoria categoria = new japedidos.produto.Categoria(id, nome, descricao);
                     c.add(categoria);
                 }
-                japedidos.produto.Categoria[] categorias = new japedidos.produto.Categoria[c.size()];
-                c.toArray(categorias);
-                return categorias;
-                
+                if (c.size() > 0) {
+                    japedidos.produto.Categoria[] categorias = new japedidos.produto.Categoria[c.size()];
+                    c.toArray(categorias);
+                    return categorias;
+                } else {
+                    return null;
+                }
             } catch (SQLException e) {
                 JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de parse", JOptionPane.ERROR_MESSAGE);
                 
             }
             
+            return null;
+        }
+    }
+
+    static public class Unidade {
+        public static final String TABLE = "unidade";
+        public static final String[] COLUMNS = {"id", "nome", "abreviacao"};
+        
+        public static int insert(japedidos.produto.Unidade u) {
+//            if (u.getId() != -1) { // Unidade não pode ser existente
+//                throw new IllegalArgumentException();
+//            }
+            if (u != null && u.getId() == -1) { // Só cadastra se não houver id
+                try {
+                    Connection conn = BD.getConnection();
+                    PreparedStatement insert = conn.prepareStatement(String.format("INSERT INTO %s(nome, abreviacao) VALUES (?, ?)", TABLE));
+                    insert.setString(1, u.getNome());
+                    insert.setString(2, u.getAbreviacao());
+                    int r = insert.executeUpdate();
+
+                    insert.close();
+                    conn.close();
+                    return r;
+                } catch (SQLException e) {
+                    JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de cadastro", JOptionPane.ERROR_MESSAGE);    
+                }
+            }
+            return 0;
+        }
+        public static int update(japedidos.produto.Unidade u) {
+            if (u != null && u.getId() != -1) { // Só atualiza se for unidade existente
+                try {
+                    Connection conn = BD.getConnection();
+                    PreparedStatement update = conn.prepareStatement(String.format("UPDATE %s SET nome = ?, abreviacao = ? WHERE id = ?", TABLE));
+                    update.setString(1, u.getNome());
+                    update.setString(2, u.getAbreviacao());
+                    update.setString(3, String.valueOf(u.getId()));
+
+                    int r = update.executeUpdate();
+
+                    update.close();
+                    conn.close();
+
+                    return r;
+                } catch(SQLException e) {
+                    JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de alteração de unidade", JOptionPane.ERROR_MESSAGE);
+                }
+            }
+            
+            return 0;
+        }
+        public static int delete(japedidos.produto.Unidade u) {
+            if (u != null && u.getId() != -1) { // Só deleta se for unidade já existente no banco
+                try {
+                    Connection conn = BD.getConnection();
+                    PreparedStatement delete = conn.prepareStatement(String.format("DELETE FROM %s WHERE id = ?", TABLE));
+                    delete.setString(1, String.valueOf(u.getId()));
+                    
+                    int r = delete.executeUpdate();
+                    
+                    delete.close();
+                    conn.close();
+                    
+                    return r;
+                } catch(SQLException e) {
+                    JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de deleção de unidade", JOptionPane.ERROR_MESSAGE);
+                }
+            }
+            return 0;
+        }
+        
+        public static japedidos.produto.Unidade selectLast() {
+            try {
+                Connection conn = BD.getConnection();
+                PreparedStatement select = conn.prepareStatement(String.format("SELECT id, nome, abreviacao FROM %s ORDER BY id DESC LIMIT 1", TABLE));
+                ResultSet rs = select.executeQuery();
+                japedidos.produto.Unidade unidade = null;
+                
+                if (rs.next()) {
+                    unidade = new japedidos.produto.Unidade(rs.getInt(1), rs.getString(2), rs.getString(3));
+                }
+                
+                select.close();
+                conn.close();
+                
+                return unidade;
+            } catch (SQLException e) {
+                JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de busca", JOptionPane.ERROR_MESSAGE);    
+            }
+            return null;
+        }
+        
+        public static japedidos.produto.Unidade[] selectAll() {
+            try {
+                Connection conn = BD.getConnection();
+                PreparedStatement select = conn.prepareStatement(String.format("SELECT id, nome, abreviacao FROM %s", TABLE));
+                ResultSet rs = select.executeQuery();
+                
+                japedidos.produto.Unidade[] unidades = parse(rs);
+                
+                select.close();
+                conn.close();
+                
+                return unidades;
+                
+            } catch (SQLException e) {
+                JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de busca", JOptionPane.ERROR_MESSAGE);    
+            }
+            return null;
+        }
+        
+        public static japedidos.produto.Unidade[] parse(ResultSet rs) {
+            try {
+                ArrayList<japedidos.produto.Unidade> u = new ArrayList<japedidos.produto.Unidade>();
+                while(rs.next()) {
+                    final int id = rs.getInt(1);
+                    final String nome = rs.getString(2);
+                    final String abreviacao = rs.getString(3);
+                    japedidos.produto.Unidade unidade = new japedidos.produto.Unidade(id, nome, abreviacao);
+                    u.add(unidade);
+                }
+                if (u.size() > 0) {
+                    japedidos.produto.Unidade[] unidades = new japedidos.produto.Unidade[u.size()];
+                    u.toArray(unidades);
+                    return unidades;                    
+                } else {
+                    return null;
+                }
+            } catch (SQLException e) {
+                JOptionPane.showMessageDialog(null, e.getMessage(), "Erro de parse", JOptionPane.ERROR_MESSAGE);
+            }
             return null;
         }
     }
